Replace the TUI's unimplemented --dry-run with safe demo mode
The flag was accepted, documented as "preview commands without executing
them", and then dropped as _dry_run without ever reaching the pipeline. It has
been that way since the TUI was first added, so a --dry-run run zapped and
repartitioned the disk exactly like a normal one: a flag that reads as safety
while providing none is worse than no flag at all.

Implementing it faithfully is not worth it either. The pipeline reaches storage
through four independent paths -- subprocesses via CommandRunner, zfskit for
every pool and dataset operation, direct filesystem writes into the target, and
libalpm for packages -- so a dry-run CommandRunner would intercept one of four
and still create the pool, write the target's configuration and install
packages. Demo mode is the project's existing answer to the same need and it
gates at the choke point instead: the wizard stays fully usable and every probe
stays real, but the install action never starts the pipeline.

Give the TUI the same semantics the Slint UI already has, and move the trigger
into core so a single definition decides what turns demo mode on for both
binaries and for the ISO's boot entry. The shared predicate no longer matches
arbitrary arguments that merely begin with the demo option's name.
